requirements.txt
# Core dependencies
pyyaml>=6.0
numpy>=1.24.0
numpy>=1.20
scipy>=1.7

# Development dependencies (install with: pip install -r requirements.txt -r requirements-dev.txt)
